+++ arm_cmsis.h
@@ -1,6 +1,6 @@
 /*!\file arm_cmsis.h
 ** \author SMFSW
-** \copyright MIT (c) 2017-2018, SMFSW
+** \copyright MIT (c) 2017-2019, SMFSW
 ** \brief ARM link with CMSIS files
 ** \warning Latest ARM chips might now be recognized; if not, define at project level:
 ** 			- for STM32 families (STMicroelectronics):