+++ arm_errors.h
@@ -1,6 +1,6 @@
 /*!\file arm_errors.h
 ** \author SMFSW
-** \copyright MIT (c) 2017-2018, SMFSW
+** \copyright MIT (c) 2017-2019, SMFSW
 ** \brief ARM user errors
 **/
 /****************************************************************/