+++ sarmfsw.h
@@ -1,6 +1,6 @@
 /*!\file sarmfsw.h
 ** \author SMFSW
-** \copyright MIT (c) 2017-2018, SMFSW
+** \copyright MIT (c) 2017-2019, SMFSW
 ** \brief sarmfsw ARM common headers
 */
 /****************************************************************/